+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/MainConfigurationTest.php
@@ -88,6 +88,55 @@ class MainConfigurationTest extends TestCase
         $this->assertEquals('a_token_id', $processedConfig['firewalls']['stub']['logout']['csrf_token_id']);
     }
 
+    public function testLogoutCsrf()
+    {
+        $config = [
+            'firewalls' => [
+                'custom_token_generator' => [
+                    'logout' => [
+                        'csrf_token_generator' => 'a_token_generator',
+                        'csrf_token_id' => 'a_token_id',
+                    ],
+                ],
+                'default_token_generator' => [
+                    'logout' => [
+                        'enable_csrf' => true,
+                        'csrf_token_id' => 'a_token_id',
+                    ],
+                ],
+                'disabled_csrf' => [
+                    'logout' => [
+                        'enable_csrf' => false,
+                    ],
+                ],
+                'empty' => [
+                    'logout' => true,
+                ],
+            ],
+        ];
+        $config = array_merge(static::$minimalConfig, $config);
+
+        $processor = new Processor();
+        $configuration = new MainConfiguration([], []);
+        $processedConfig = $processor->processConfiguration($configuration, [$config]);
+
+        $assertions = [
+            'custom_token_generator' => [true, 'a_token_generator'],
+            'default_token_generator' => [true, 'security.csrf.token_generator'],
+            'disabled_csrf' => [false, null],
+            'empty' => [false, null],
+        ];
+        foreach ($assertions as $firewallName => [$enabled, $tokenGenerator]) {
+            $this->assertEquals($enabled, $processedConfig['firewalls'][$firewallName]['logout']['enable_csrf']);
+            if ($tokenGenerator) {
+                $this->assertEquals($tokenGenerator, $processedConfig['firewalls'][$firewallName]['logout']['csrf_token_generator']);
+                $this->assertEquals('a_token_id', $processedConfig['firewalls'][$firewallName]['logout']['csrf_token_id']);
+            } else {
+                $this->assertArrayNotHasKey('csrf_token_generator', $processedConfig['firewalls'][$firewallName]['logout']);
+            }
+        }
+    }
+
     public function testDefaultUserCheckers()
     {
         $processor = new Processor();